+++ jetty/src/main/scala/org/http4s/server/jetty/JettyBuilder.scala
@@ -48,6 +48,32 @@ sealed class JettyBuilder[F[_]] private (
 
   private[this] val logger = getLogger
 
+  @deprecated("Retained for binary compatibility", "0.20.23")
+  private[JettyBuilder] def this(
+      socketAddress: InetSocketAddress,
+      threadPool: ThreadPool,
+      idleTimeout: Duration,
+      asyncTimeout: Duration,
+      shutdownTimeout: Duration,
+      servletIo: ServletIo[F],
+      sslConfig: SslConfig,
+      mounts: Vector[Mount[F]],
+      serviceErrorHandler: ServiceErrorHandler[F],
+      banner: immutable.Seq[String]
+  )(implicit F: ConcurrentEffect[F]) = this(
+    socketAddress = socketAddress,
+    threadPool = threadPool,
+    idleTimeout = idleTimeout,
+    asyncTimeout = asyncTimeout,
+    shutdownTimeout = shutdownTimeout,
+    servletIo = servletIo,
+    sslConfig = sslConfig,
+    mounts = mounts,
+    serviceErrorHandler = serviceErrorHandler,
+    supportHttp2 = false,
+    banner = banner
+  )
+
   private def copy(
       socketAddress: InetSocketAddress = socketAddress,
       threadPool: ThreadPool = threadPool,